chore: moves codelens and folding to setupMonaco
