--- src/util/get-reference-section.ts
@@ -1,15 +1,0 @@
-const extruderRegExp = /^extruder[0-9]+$/
-
-const getReferenceSection = (sectionName: string) => {
-  if (sectionName.startsWith('stepper_')) {
-    return 'stepper'
-  }
-
-  if (extruderRegExp.test(sectionName)) {
-    return 'extruder'
-  }
-
-  return sectionName
-}
-
-export default getReferenceSection
