+++ src/util/get-vue-app.ts
@@ -1,0 +1,5 @@
+const getVueApp = () => {
+  return (document.querySelector('#app') as any).__vue__
+}
+
+export default getVueApp
